requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
requests==2.31.0
joblib==1.3.2
numpy==1.23.5
scikit-learn==1.1.3
tensorflow==2.11.0
pillow==9.5.0
opencv-python==4.7.0.72
